JSON equipment properties import
Started implementing the JSON equipment import. Now creates a list of dicts to represent the equipment. Need to change the way that the simulation iterates through the equipment.  Right now we are tracking a 'blob' and pushing it through the entire chain in each loop. This isn't particularly accurate. There should be some concept of transit time (maybe intermediate dwell objects?).